requirements.txt
autopep8==1.4.4
bcrypt==3.1.7
cffi==1.13.2
cryptography==2.8
paramiko==2.7.1
pycodestyle==2.5.0
pycparser==2.19
PyNaCl==1.3.0
python-dotenv==0.10.3
scp==0.13.2
six==1.13.0
